Fix Muon's zeropower default steps & update function to current version
Hey! Looks like the zeropower default steps were super low (2), this can cause convergence issues as I understand it so I updated it to 5 which seems good in a lot of usecases.

Inner bfloat16 seems to be a big win for speed, considering how much the vectors are moving the precision seems to have negligible effects (even at larger scales).

Did update so that the output cast is in the original dtype for comparability purposes.